Docker plugin - add tag "latest" when tag not specified
When there is no tag specified, the entire docker repo
is downloaded. However only the one tagged as "latest" is ran.
So in order to avoid downloading the entire repo,
if user didn't specify any tag, append "latest" to download
only the image tagged with latest.
See related issue and discussion here
https://github.com/dotcloud/docker/issues/5047

Change-Id: I5e875733797e8fcebd557c35e230dcdb6fb5512e
Signed-off-by: Alissa Bonas <abonas@redhat.com>