Cockpit UI Plugin for webadmin
With this plugin, the host main-tab shows new subtab embedding Cockpit
which is running on the selected host.

The host's right-click menu is enriched for new 'Cockpit' action leading
to opening new browser window/tab with the host's Cockpit.

When Cockpit is not running on the selected host, the menu action is disabled
and the embedding subtab shows content of noCockpit.html page.

See the README for more information.
See cockpit.json for configuration.

Change-Id: If70347aadcb30c14a0db05d9e9c4398cba09a05b
Signed-off-by: Marek Libra <mlibra@redhat.com>